Fix bug, Move query to next line, simplify code
I added a newline character to the download query because it's more convenient to not have to scroll to the end of a window just to see that you're supposed to type 'y'.

There was an unnoticed bug in the `cont = args.yes or etc etc etc or "y"` line in that `"y"` is a non-empty string, so it will always return true (Since the input was not checked for equality to the desired strings it would always return True as well). This means that no matter what the user entered, the Boolean `cont` would always evaluate to `True`. I fixed this and trimmed up the `if` statement that is now on line 46, replacing the unnecessary `if cont in [True, "y", "Y", "yes", ""]:` statement (Since `cont` is already a Boolean) with simply `if cont:`

Other than that I'm enjoying the program and may start using it in larger projects or for my own convenience.